requirements.txt
# Function dependencies, for example:
# package>=version

google-resumable-media==0.6.0
google-cloud-storage==1.30.0
google-cloud-bigquery==1.26.1
pandas
numpy==1.18.1
scikit-learn==0.22
